モジュールの統合
