+++ lib/Meridian/Core/MeridianPlugin/include/mrd_module/sv_dxl2.h
@@ -5,11 +5,13 @@
 #include "main.h"
 
 //==================================================================================================
-//  DYNOMIXELサーボ関連の処理
+//  DYNAMIXELサーボ関連の処理
 //==================================================================================================
 
-void mrd_servo_dxl20_x() {
-  //
+/// @brief スタブ関数.
+/// @return 常にfalseを返す.
+bool mrd_servo_dxl20_x() {
+  return false;
 }
 
 #endif // __MERIDIAN_SERVO_DYNAMIXEL_H__
