+++ lib/Meridian/Core/MeridianPlugin/include/mrd_module/sv_ftbrx.h
@@ -1,17 +1,17 @@
-#ifndef __MERIDIAN_SERVO_DYNAMIXEL_H__
-#define __MERIDIAN_SERVO_DYNAMIXEL_H__
+#ifndef __MERIDIAN_SERVO_FUTABA_RSxTTL_H__
+#define __MERIDIAN_SERVO_FUTABA_RSxTTL_H__
 
 #include "config.h"
 #include "main.h"
 
 //==================================================================================================
-//  DYNAMIXELサーボ関連の処理  ----------------------------------------------------------------------
+//  FUTABA RSxTTLサーボ関連の処理
 //==================================================================================================
 
 /// @brief スタブ関数.
 /// @return 常にfalseを返す.
-bool mrd_servo_dxl20_x() {
+bool mrd_servo_ftbrs_x() {
   return false;
 }
 
-#endif // __MERIDIAN_SERVO_DYNAMIXEL_H__
+#endif // __MERIDIAN_SERVO_FUTABA_RSxTTL_H__
