+++ lib/Meridian/Core/MeridianPlugin/include/mrd_module/sv_ftc.h
@@ -5,7 +5,7 @@
 #include "main.h"
 
 //==================================================================================================
-//  FEETECH TTLサーボ関連の処理  --------------------------------------------------------------------
+//  FEETECH TTLサーボ関連の処理
 //==================================================================================================
 
 /// @brief スタブ関数.
@@ -14,4 +14,4 @@ bool mrd_servo_ftc_x() {
   return false;
 }
 
-#endif // __MERIDIAN_SERVO_FUTABA_RSxTTL_H__
+#endif // __MERIDIAN_SERVO_FEETECH_H__
