--- lib/Meridian/ESP32/MeridianPlugin_for_Arudino/src/mrd_module/sv_ftbrx.h
@@ -1,17 +1,0 @@
-#ifndef __MERIDIAN_SERVO_FUTABA_RSxTTL_H__
-#define __MERIDIAN_SERVO_FUTABA_RSxTTL_H__
-
-#include "config.h"
-#include "main.h"
-
-//==================================================================================================
-//  FUTABA RSxTTLサーボ関連の処理  ------------------------------------------------------------------
-//==================================================================================================
-
-/// @brief スタブ関数.
-/// @return 常にfalseを返す.
-bool mrd_servo_ftbrs_x() {
-  return false;
-}
-
-#endif // __MERIDIAN_SERVO_FUTABA_RSxTTL_H__
