--- lib/Meridian/Teensy40/MeridianPlugin_for_Arudino/src/mrd_module/sv_ftbrx.h
@@ -1,15 +1,0 @@
-#ifndef __MERIDIAN_SERVO_FUTABA_RSxTTL_H__
-#define __MERIDIAN_SERVO_FUTABA_RSxTTL_H__
-
-#include "config.h"
-#include "main.h"
-
-//==================================================================================================
-//  FUTABA RSxTTLサーボ関連の処理
-//==================================================================================================
-
-void mrd_servo_ftbrs_x() {
-  //
-}
-
-#endif // __MERIDIAN_SERVO_FUTABA_RSxTTL_H__
