--- lib/Meridian/Teensy40/MeridianPlugin_for_Arudino/src/mrd_module/sv_ftc.h
@@ -1,15 +1,0 @@
-#ifndef __MERIDIAN_SERVO_FEETECH_H__
-#define __MERIDIAN_SERVO_FEETECH_H__
-
-#include "config.h"
-#include "main.h"
-
-//==================================================================================================
-//  FEETECH TTLサーボ関連の処理
-//==================================================================================================
-
-void mrd_servo_ftc_x() {
-  //
-}
-
-#endif // __MERIDIAN_SERVO_FEETECH_H__
